+++ ScrumLearning/ScrumLearning/src/main/java/uppgift/Statistics/Player.java
@@ -1,20 +1,30 @@
 package uppgift.Statistics;
 
-import uppgift.user.Users;
+import java.util.HashSet;
+import java.util.Set;
 
 public class Player {
-    private Users user;
+    private String username;
+    private static Set<String> registeredUsernames = new HashSet<>();
     private StatisticManager statisticManager;
     private StatisticSaver statisticSaver;
 
     public Player(String username, StatisticSaver statisticSaver) {
-        this.user = new Users(username);
+        if (isUsernameTaken(username)) {
+            throw new IllegalArgumentException("Username already taken.");
+        }
+        this.username = username;
+        registeredUsernames.add(username);
         this.statisticManager = new StatisticManager();
         this.statisticSaver = statisticSaver;
     }
 
+    public static boolean isUsernameTaken(String username) {
+        return registeredUsernames.contains(username);
+    }
+
     public String getUsername() {
-        return user.getUsername();
+        return username;
     }
 
     public StatisticManager getStatisticManager() {
